+++ src/styles/pages/search.module.scss
@@ -1,4 +1,5 @@
 @import 'styles/uswdsDependencies';
+@import 'styles/sfds/sfdsDependencies';
 
 .pageTitle {
   margin-top: units(6);
@@ -26,4 +27,31 @@
 
 .noResults {
   display: flex;
+}
+
+.searchFilter {
+  background: var(--background-fill);
+  border: 1px solid $light-gray;
+  border-radius: 8px;
+  padding: units(2);
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
+
+  > h3 {
+    margin: 0;
+    font-weight: 500;
+  }
+
+  > .headerContainer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
+
+.checkbox {
+  background-color: transparent;
+}
+
+.buttonContainer {
+  margin-top: 12px;
 }
